requirements.txt
Flask==1.0.2
Pillow==5.3.0
